+++ packages/constants/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./providers/index.js";
 export * from "./seed.js";
 export * from "./version.js";
+export * from "./legal.js";